+++ src/lib/desktop-path.ts
@@ -1,0 +1,15 @@
+import { existsSync } from 'fs';
+import os from 'os';
+import path from 'node:path';
+
+/**
+ * Get the path to the desktop directory
+ *
+ * @returns {string}
+ */
+export function getDesktopPath(): string {
+  const homeDir = path.resolve(os.homedir());
+  const desktopDir = path.resolve(os.homedir(), 'Desktop');
+
+  return existsSync(desktopDir) ? desktopDir : homeDir;
+}
