+++ src/lib/index.ts
@@ -1,0 +1,5 @@
+export * from './desktop-path';
+export * from './utils';
+export * from './window-manager';
+export * from './ffmpeg';
+export * from './ipc-handlers';
