+++ src/lib/utils.ts
@@ -1,6 +1,13 @@
-import { type ClassValue, clsx } from 'clsx'
-import { twMerge } from 'tailwind-merge'
+import { type ClassValue, clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
 
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
+/**
+ * A utility function to merge Tailwind CSS classes with the `clsx` utility.
+ *
+ * @param {ClassValue[]} inputs
+ *
+ * @returns {string}
+ */
+export function cn(...inputs: ClassValue[]): string {
+  return twMerge(clsx(inputs));
 }
